requirements.txt
opencv-python
pytesseract
pydirectinput
PyQt6
numpy
Pillow
pyautogui
